exp.host server
The current version of the server running on exp.host with deployment scripts. Our code base uses io.js and ES7 async/await in conjunction with koa, a generator-based server.

The database is rethinkdb which has a good query API that is an embedded DSL. More info to come on how to set this up for development against a testing DB.

The server runs in pm2 which provides stats about the process and keeps it running.

We use gulp for tooling. Run `npm install -g gulp`. Also the best way to get io.js is through nvm. Currently we're using io.js 1.8.1 but this server should run fine on io.js 2.0.x.