[Web] Server-side rendering on! Integrated assets-module-webpack-plugin
We can now currently serve HTML rendered from the server with the help of assets-module-webpack-plugin, which generates stub JS modules for our LESS/CSS and image files so that they can be required on the server without Node's crashing.

This lets us do server-side rendering without needing to build a webpack bundle for Node. Will deploy this so the site is faster.